qpqpqp

+++ src/Main.java
@@ -1,10 +1,10 @@
 public class Main {
 	public static void main(String[] args) {
 
-		int num;
-		
+		double num;
+
 		num = 1;
-		
+
 		System.out.println(num);
 	}
 }
